Updated project:
- added filter by year and realised an ability to seeing value in console
