+++ src/App.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import './App.css';
-import Expenses from "./components/Expensses/Expenses";
+import Expenses from "./components/Expenses/Expenses";
 import NewExpense from "./components/NewExpense/NewExpense";
+
 
 const App = () => {
     const expenses = [
